// src/common/enums/order-field.enum.ts
export enum OrderFieldEnum {
  ID = 'id',
  NAME = 'name',
  SURNAME = 'surname',
  EMAIL = 'email',
  PHONE = 'phone',
  AGE = 'age',
  COURSE = 'course',
  COURSE_FORMAT = 'course_format',
  COURSE_TYPE = 'course_type',
  SUM = 'sum',
  ALREADYPAID = 'alreadyPaid',
  CREATED_AT = 'created_at',
  UTM = 'utm',
  MSG = 'msg',
  STATUS = 'status',
  GROUP = 'group',
  MANAGER = 'manager',
}